requirements.txt
bpy
pillow
OpenEXR
argparse
opencv-python